requirements.txt
openai
openai-whisper
ipykernel
pyyaml